+++ Source/ASVideoPlayerNode.mm
@@ -26,6 +26,7 @@
 #import <AsyncDisplayKit/AsyncDisplayKit.h>
 #import <AsyncDisplayKit/ASDefaultPlaybackButton.h>
 #import <AsyncDisplayKit/ASDisplayNode+Subclasses.h>
+#import <AsyncDisplayKit/ASDisplayNodeInternal.h>
 #import <AsyncDisplayKit/ASThread.h>
 
 static void *ASVideoPlayerNodeContext = &ASVideoPlayerNodeContext;
@@ -334,6 +335,8 @@ static void *ASVideoPlayerNodeContext = &ASVideoPlayerNodeContext;
 
 - (void)_locked_createPlaybackButton
 {
+  ASAssertLocked(__instanceLock__);
+  
   if (_playbackButtonNode == nil) {
     _playbackButtonNode = [[ASDefaultPlaybackButton alloc] init];
     _playbackButtonNode.style.preferredSize = CGSizeMake(16.0, 22.0);
@@ -357,6 +360,8 @@ static void *ASVideoPlayerNodeContext = &ASVideoPlayerNodeContext;
 
 - (void)_locked_createFullScreenButton
 {
+  ASAssertLocked(__instanceLock__);
+  
   if (_fullScreenButtonNode == nil) {
     _fullScreenButtonNode = [[ASButtonNode alloc] init];
     _fullScreenButtonNode.style.preferredSize = CGSizeMake(16.0, 22.0);
@@ -374,6 +379,8 @@ static void *ASVideoPlayerNodeContext = &ASVideoPlayerNodeContext;
 
 - (void)_locked_createElapsedTextField
 {
+  ASAssertLocked(__instanceLock__);
+  
   if (_elapsedTextNode == nil) {
     _elapsedTextNode = [[ASTextNode alloc] init];
     _elapsedTextNode.attributedText = [self timeLabelAttributedStringForString:@"00:00"
@@ -387,6 +394,8 @@ static void *ASVideoPlayerNodeContext = &ASVideoPlayerNodeContext;
 
 - (void)_locked_createDurationTextField
 {
+  ASAssertLocked(__instanceLock__);
+  
   if (_durationTextNode == nil) {
     _durationTextNode = [[ASTextNode alloc] init];
     _durationTextNode.attributedText = [self timeLabelAttributedStringForString:@"00:00"
@@ -401,6 +410,8 @@ static void *ASVideoPlayerNodeContext = &ASVideoPlayerNodeContext;
 
 - (void)_locked_createScrubber
 {
+  ASAssertLocked(__instanceLock__);
+  
   if (_scrubberNode == nil) {
     __weak __typeof__(self) weakSelf = self;
     _scrubberNode = [[ASDisplayNode alloc] initWithViewBlock:^UIView * _Nonnull {
@@ -445,6 +456,8 @@ static void *ASVideoPlayerNodeContext = &ASVideoPlayerNodeContext;
 
 - (void)_locked_createControlFlexGrowSpacer
 {
+  ASAssertLocked(__instanceLock__);
+  
   if (_controlFlexGrowSpacerSpec == nil) {
     _controlFlexGrowSpacerSpec = [[ASStackLayoutSpec alloc] init];
     _controlFlexGrowSpacerSpec.style.flexGrow = 1.0;